+++ core/src/main/java/uk/ac/bbsrc/tgac/miso/core/data/Library.java
@@ -339,8 +339,8 @@ public interface Library
 
   public IndexFamily getCurrentFamily();
 
-  Long getDnaSize();
+  Integer getDnaSize();
 
-  void setDnaSize(Long dnaSize);
+  void setDnaSize(Integer dnaSize);
 
 }
